fix: comprehensive code review fixes for security, type alignment and robustness
Security:
- Remove hardcoded SECRET_KEY/DATABASE_URL defaults, force env vars
- Add BACKEND_CORS_ORIGINS config, replace wildcard CORS
- Add repair status transition validation (state machine)
- Validate dorm_id ownership for repair creation
- Add admin self-delete/self-disable protection
- Show test credentials only in dev mode
- Add password min length (6 chars) and role enum constraints
- Validate assigned_to user exists and is repairman role

Type alignment (frontend ↔ backend):
- Rewrite api.ts types to match backend schemas exactly
- Update all page components (Dashboard, Dormitories, Students, Repairs, Charges, Announcements) field names
- Fix UserRole: dorm_manager → repairman, User.name → real_name
- Fix charge.amount Decimal→float serialization (was causing white screen)

Auth & robustness:
- Verify token validity on app init via getCurrentUser()
- Fix 401 interceptor: add concurrency guard, carry redirect param
- Use env var for API baseURL, fix hardcoded localhost
- Fix Login error message: read ApiError.message instead of always showing generic
- Add login redirect back to original page
- Use React useId() instead of Math.random() for Input IDs

Deprecation & consistency:
- Replace on_event with lifespan context manager
- Replace datetime.utcnow() with datetime.now(timezone.utc)
- Replace declarative_base() with DeclarativeBase class
- Register endpoint returns 501 instead of 200
- Fix statistics None→0 for all aggregate functions
- Add CheckInRecord writes in check_in/check_out
- Remove duplicate AuthProvider in App.tsx

+++ app/schemas/charge_schema.py
@@ -1,13 +1,12 @@
 from pydantic import BaseModel
 from typing import Optional
 from datetime import datetime, date
-from decimal import Decimal
 
 
 class ChargeBase(BaseModel):
     """费用账单基础模型"""
     charge_type: str
-    amount: Decimal
+    amount: float
     charge_date: date
     due_date: Optional[date] = None
     memo: Optional[str] = None
